requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
groq==0.8.0
python-dotenv==1.0.1
httpx==0.27.0
joblib==1.4.2
sse-starlette==2.1.0
